Initial default file upload.
Initial commit of the blank game template generated by UE5 for use with the Social Chat Environment.

Whilst blank with no assets was specified, UE5 has created an initial level with a streaming default landscape, along sky and lighting etc.

The default settings and content need to be reviewed and either changed, ammended or deleted as appropriate.